+++ ChangeLog.md
@@ -2,8 +2,33 @@ ISAAC Changelog
 
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers where provided, and the git commit history.
 
-* 2017/09/?? - 5.03 - PENDING
-    *
+* 2017/09/?? - 6.00 - PENDING
+    * Fix a documentation bug on the required format for a VHAT XML source file version number, add missing regular expression to enforce the rule.
+
+* 2017/09/15 - 5.08 - R5 build v3
+    * Fix a commit lock bug that hangs the server to any write request when it gets triggered.
+    * Fix a bug on description remove in vhat delta import that did a double edit on the extended description types, leading to db model warnings.
+    
+* 2017/09/13 - 5.07 - R5 build v2
+    * Fix some critical errors in the XML Delta import dealing with designations.
+
+* 2017/09/12 - 5.06
+    * Updates to VetsExporter from inconsistencies found during testing
+    * Updates to Frills.getDescriptionExtendedTypeConcept to allow returning inactive descriptions
+    * Documentation updates
+    * R5 release build 
+
+* 2017/09/08 - 5.05
+    * Fixes to trim XML import data of Properties, Designations, and supported trimming new Subsets and Types (Property and Relationship) (585147)
+    * Fixes to move SubsetMemberships during MoveFromConceptCode (583082)
+
+* 2017/09/07 - 5.04
+    * Fixes for trimming XML input data 
+    * Changes to support MoveFromConceptCode via XML import (583082)
+
+* 2017/09/06 - 5.03
+    * Fix for UPDATEs to Relationships where NewTargetCode may not be present (Mike's has_parent test cases) (583864)
+    * Changes to modify all nested sememes of a Designation to be inactivated when the parent Designation is 'removed' (583063)
 
 * 2017/09/01 - 5.02
     * Changes to delta importer to explicitely cause REMOVE directives in data elements to set active=false (583063)